Changed format for data returned from parsers
A meta-meta dictionary is now returned. The original metadata
(information directly from the source) is now provided in a
sub-dictionary under the "meta" key. Information that we have
interpreted based on this metadata is stored in the top level.

For the OCSO parser, the original call number is now preserved. A new
field has been added to all calls that is to be used as a unique
identifier. This replaces the getHash() function.

Instead of reporting all of the data for expired calls, now only the
unique key is reported.