Add database calls to web ingest script
New calls are added to the database and expired calls are marked as
such. There is no attempt to avoid duplicate data.

The web portion of the project now has a config file, located in
Web\config.php. Right now it just stores the database login details.
This might change.