Rework config sections
There are now 2 sections, 'main' and 'paths'. Main contains general
program settings. Paths contains URLs and file locations.

Some new settings were also added to determine if the client should
connect to a remote server and the locations of that servers resources.

When the program loads, if no config exists, a default one will be
created and saved.