Add geocoding of location values
Parse scripts now create a geocode-friendly of the address and send it
to a geocoding API to return additional information, including latitude
and logitude. This information is stored in a separate table in the
database, geocodes. Entries in this table are linked to the calls
through numeric IDs.

Clients retrieve new geocode requests whenever they submit new calls to
the server. They do not geocode locations on their own, only when the
server gives it to them.