Add generic replacement system for source parsers
This adds a new variable passed into the parser scripts, replacements.
The script can fill this dictionary with find->replace pairs and the
client will check each element (or word) of the returned locations and
make the appropriate replacements. An element will only be replaced if
it matches FULLY. It is not an inline find/replace.