Handle some geocode reporting errors
Requests are now cached in the client. The client will not request
geocoding for requests that have already been resolved. If a report
fails, the client will try again with the same data later. A copy of the
data will also be saved locally for analysis.

On the server side, some array key errors were fixed.