[Web] Normalize call times shown in /list for sorting
Adds a time_format column to the sources table which specifies the format that the source gives for call times. The times given by the source are merged with the time the call was found should the date not be provided. This should enable sorting of the call list by the call time.

Also removes old and unused timezone related code.