Convert to Google Maps JavaScript API
This replaces the previous Google Visualization API for the maps. The new system is a lot more flexibile and allows markers to be added and removed whenever without redrawing the whole map. This API requires a key, which can be obtained from the Google Cloud API dashboard (https://console.cloud.google.com/apis/dashboard).